Task_mangement_AWS_CDK/requirements.txt
attrs==25.3.0
aws-cdk-lib==2.190.0
aws-cdk.asset-awscli-v1==2.2.235
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==41.2.0
black==25.1.0
cattrs==24.1.3
click==8.1.8
constructs==10.4.2
coverage==7.8.0
flake8==7.2.0
importlib_resources==6.5.2
iniconfig==2.1.0
isort==6.0.1
jsii==1.111.0
mccabe==0.7.0
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
publication==0.0.3
pycodestyle==2.13.0
pyflakes==3.3.2
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.13.2

boto3~=1.38.13
botocore~=1.38.13